Remove embedding fonts.  Add ThickLine
Since QB64PE v3.10+ now comes with the _EMBEDDED commands, these routines are redundant and no longer necessary and have been removed to simplify the overall size and length of the library.

SUB ThickLine has been added to allow for ease in creating lines of various thickness in a program.