Update to remove unnecessary routines now native to QB64PE v 4.1+
Offsets no longer have the same restrictions in QB64PE so there's no need for ConvertOffset.

Base64 conversions are now native to QB64PE, so we no longer need to do them ourselves.  This removes the routines which converted to and from those bases.

Changed the GetFileDir routine to make use of the _FILES command and remove direntry.h requirments.  Separated the command to get file and directory separately from now on.